First attempt at restructuring the module.
